Update Solin Twiner, enrol user event

+++ templates/list.php
@@ -4,41 +4,78 @@
  * list template
  */
 $notify_remove = optional_param('notify_remove', 0, PARAM_INT);
-if ($notify_remove)
+$enrol_remove = optional_param('enrol_remove', 0, PARAM_INT);
+if ($notify_remove || $enrol_remove)
 {
-    $DB->delete_records('twiner_notifications', array('id' => $notify_remove));
+    if ($notify_remove) $DB->delete_records('twiner_notifications', array('id' => $notify_remove));
+    if ($enrol_remove) $DB->delete_records('twiner_enrol', array('id' => $enrol_remove));
 	echo "<p>" . get_string('trigger_removed', 'local_solin_twiner') . "</p>\n";
 }
 
-$sql = '
-SELECT t1.id, t1.creator_id, t1.target_id, t1.subject, t2.eventname FROM {twiner_notifications} AS t1
-    LEFT JOIN {twiner_triggers} AS t2 ON t1.trigger_id = t2.id';
-$notifications = $DB->get_records_sql($sql);
+$sql_notifications  = "SELECT tn.id, tn.creator_id, tn.target_id, tn.subject, tt.eventname ";
+$sql_notifications .= "FROM {twiner_notifications} AS tn ";
+$sql_notifications .= "LEFT JOIN {twiner_triggers} AS tt ON tn.trigger_id = tt.id ";
+$notifications = $DB->get_records_sql($sql_notifications);
 
-if ($notifications)
+$sql_enrollments  = "SELECT te.id, te.user_id, te.course_id, tt.eventname, c.fullname ";
+$sql_enrollments .= "FROM {twiner_enrol} te ";
+$sql_enrollments .= "JOIN {twiner_triggers} AS tt ON te.trigger_id = tt.id ";
+$sql_enrollments .= "JOIN {course} AS c ON te.course_id = c.id ";
+$enrollments = $DB->get_records_sql($sql_enrollments);
+
+if ($notifications || $enrollments)
 {
-	?>
-	<h2><?=get_string('notifications','local_solin_twiner');?></h2>
+	if ($notifications)
+	{
+		?>
+		<h2><?= get_string('notifications', 'local_solin_twiner'); ?></h2>
 		<table class="generaltable">
-			<tr style="font-weight: bold;">
-				<td><?=get_string('event','local_solin_twiner');?></td>
-				<td><?=get_string('type','local_solin_twiner');?></td>
-				<td><?=get_string('target','local_solin_twiner');?></td>
-	            <td><?=get_string('subject','local_solin_twiner');?></td>
-				<td>&nbsp;</td>
+			<tr>
+				<th><?= get_string('event', 'local_solin_twiner'); ?></th>
+				<th><?= get_string('type', 'local_solin_twiner'); ?></th>
+				<th><?= get_string('user', 'local_solin_twiner'); ?></th>
+				<th><?= get_string('subject', 'local_solin_twiner'); ?></th>
+				<th>&nbsp;</th>
 			</tr>
 		<? foreach ($notifications as $notification):
-		$user = $DB->get_record('user',array('id'=>$notification->target_id));?>
+			$user = $DB->get_record('user', array('id' => $notification->target_id)); ?>
 			<tr>
-				<td><?=call_user_func(array($notification->eventname, 'get_name'));?></td>
-				<td><?=get_string('individual','local_solin_twiner');?></td>
-				<td><?=$user->firstname.' '.$user->lastname;?></td>
-	            <td><?= $notification->subject; ?></td>
-				<td><?=html_writer::link($twiner_url.'?type=2&notify_remove='.$notification->id,get_string('remove','local_solin_twiner'));?></td>
+				<td><?= call_user_func(array($notification->eventname, 'get_name')); ?></td>
+				<?php // fixme ?>
+				<td><?= get_string('individual', 'local_solin_twiner'); ?></td>
+				<td><?= fullname($user); ?></td>
+				<td><?= $notification->subject; ?></td>
+				<td><?= html_writer::link($twiner_url . '?type=2&notify_remove=' . $notification->id, get_string('remove', 'local_solin_twiner')); ?></td>
 			</tr>
 		<? endforeach;?>
 		</table>
-<?php
+		<p>&nbsp;</p>
+		<?php
+	}
+	
+	if ($enrollments)
+	{
+		?>
+		<h2><?= get_string('enrollments', 'local_solin_twiner'); ?></h2>
+		<table class="generaltable">
+			<tr>
+				<th><?= get_string('event', 'local_solin_twiner'); ?></th>
+				<th><?= get_string('user', 'local_solin_twiner'); ?></th>
+				<th><?= get_string('course', 'local_solin_twiner'); ?></th>
+				<th>&nbsp;</th>
+			</tr>
+		<? foreach ($enrollments as $enrollment):
+			$user = $DB->get_record('user', array('id' => $enrollment->user_id)); ?>
+			<tr>
+				<td><?= call_user_func(array($enrollment->eventname, 'get_name')); ?></td>
+				<td><?= fullname($user); ?></td>
+				<td><?= $enrollment->fullname; ?></td>
+				<td><?= html_writer::link($twiner_url . '?type=2&enrol_remove=' . $enrollment->id, get_string('remove', 'local_solin_twiner')); ?></td>
+			</tr>
+		<? endforeach;?>
+		</table>
+		<?php
+	}
 }
 else
 {
